requirements.txt
cffi==1.15.0
click==8.1.3
cowpy==1.0.3
cryptography==35.0.0
Flask==2.1.2
Flask-Cors==3.0.10
importlib-metadata==4.11.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
neo4j==4.4.4
pycparser==2.21
PyJWT==2.3.0
pytz==2022.1
six==1.16.0
Werkzeug==2.1.2
zipp==3.8.0
